+++ udata_recommendations/__init__.py
@@ -2,5 +2,5 @@
 uData recommendations
 '''
 
-__version__ = '2.0.0.dev'
+__version__ = '2.0.0'
 __description__ = 'uData content recommendations bridge'