+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Current (in progress)
+## 2.0.0 (2020-03-11)
 
 - udata 2.0 / Python 3 support [#95](https://github.com/opendatateam/udata-recommendations/pull/95)
 - Support new hooks format [#96](https://github.com/opendatateam/udata-recommendations/pull/96)